refactor: move SavedGame static file methods' tests to separate test class
